Move get_glean_event
