+++ privaterelay/tests/utils.py
@@ -1,4 +1,13 @@
-def log_extra(log_record):
+"""Helper functions for tests"""
+
+from logging import LogRecord
+from typing import Any
+import json
+
+import pytest
+
+
+def log_extra(log_record: LogRecord) -> dict[str, Any]:
     """Reconstruct the "extra" argument to the log call"""
     omit_log_record_keys = set(
         (
@@ -30,3 +39,13 @@ def log_extra(log_record):
         for key, val in log_record.__dict__.items()
         if key not in omit_log_record_keys
     }
+
+
+def get_glean_event(caplog: pytest.LogCaptureFixture) -> dict[str, Any] | None:
+    """Return the event payload from a Glean server event log."""
+    event = None
+    for record in caplog.records:
+        if record.msg == "glean-server-event":
+            assert hasattr(record, "payload")
+            event = json.loads(record.payload)["events"][0]
+    return event
